reorganize upload_failure code
