+++ tests/unit/backend/corpora/common/utils/test_slack.py
@@ -1,7 +1,7 @@
 import json
 import os
 
-from backend.corpora.lambdas.upload_failures.slack import format_dataset_processing_failure_slack_message
+from backend.corpora.common.utils.slack import format_dataset_processing_failure_slack_message
 from tests.unit.backend.fixtures.data_portal_test_case import DataPortalTestCase
